cherry pick BPP211 migration to master

+++ Console/BExIS.Web.Shell/Areas/DCM/Controllers/ImportMetadataStructureReadSourceController.cs
@@ -1,18 +1,16 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Web;
-using System.Web.Mvc;
-using System.Web.Routing;
-using BExIS.Dcm.ImportMetadataStructureWizard;
+﻿using BExIS.Dcm.ImportMetadataStructureWizard;
 using BExIS.Dcm.Wizard;
 using BExIS.Dlm.Services.MetadataStructure;
 using BExIS.IO.Transform.Validation.Exceptions;
-using BExIS.Web.Shell.Areas.DCM.Models.ImportMetadata;
+using BExIS.Modules.Dcm.UI.Models.ImportMetadata;
+using BExIS.Utils.Helpers;
 using BExIS.Xml.Helpers.Mapping;
+using System;
+using System.Linq;
+using System.Web.Mvc;
+using System.Web.Routing;
 
-namespace BExIS.Web.Shell.Areas.DCM.Controllers
+namespace BExIS.Modules.Dcm.UI.Controllers
 {
     public class ImportMetadataStructureReadSourceController : Controller
     {
@@ -34,6 +32,9 @@ namespace BExIS.Web.Shell.Areas.DCM.Controllers
             }
 
             ReadSourceModel model = new ReadSourceModel(TaskManager.Current());
+
+            if (TaskManager.Bus.ContainsKey(ImportMetadataStructureTaskManager.IS_GENERATE))
+                model.IsGenerated = (bool)TaskManager.Bus[ImportMetadataStructureTaskManager.IS_GENERATE];
 
             if (TaskManager.Bus.ContainsKey(ImportMetadataStructureTaskManager.ROOT_NODE))
                 model.RootNode = TaskManager.Bus[ImportMetadataStructureTaskManager.ROOT_NODE].ToString();
@@ -57,7 +58,10 @@ namespace BExIS.Web.Shell.Areas.DCM.Controllers
                 TaskManager.Current().SetStatus(StepStatus.success);
             }
             else
+            {
                 ModelState.AddModelError("", "Please click generate button.");
+            }
+
 
             if (TaskManager.Current().IsValid())
             {
@@ -75,19 +79,36 @@ namespace BExIS.Web.Shell.Areas.DCM.Controllers
 
         public ActionResult SetRootNode(string name)
         {
-             TaskManager = (ImportMetadataStructureTaskManager)Session["TaskManager"];
+            TaskManager = (ImportMetadataStructureTaskManager)Session["TaskManager"];
 
-             if (TaskManager.Bus.ContainsKey(ImportMetadataStructureTaskManager.ROOT_NODE))
-                 TaskManager.Bus[ImportMetadataStructureTaskManager.ROOT_NODE] = name;
-             else
-                 TaskManager.Bus.Add(ImportMetadataStructureTaskManager.ROOT_NODE, name);
+            if (TaskManager.Bus.ContainsKey(ImportMetadataStructureTaskManager.ROOT_NODE))
+                TaskManager.Bus[ImportMetadataStructureTaskManager.ROOT_NODE] = name;
+            else
+                TaskManager.Bus.Add(ImportMetadataStructureTaskManager.ROOT_NODE, name);
 
 
-             return Content("");
+            return Content("");
         }
 
-        public ActionResult SetSchemaName(string name)
+        public JsonResult SetSchemaName(string name)
         {
+
+            if (String.IsNullOrEmpty(name))
+            {
+                return Json("A Metadata structure must have a name.", JsonRequestBehavior.AllowGet);
+            }
+
+            if (!RegExHelper.IsFilenameValid(name))
+            {
+                return Json("Name : \" " + name + " \" is invalid. These special characters are not allowed : \\/:*?\"<>|", JsonRequestBehavior.AllowGet);
+            }
+
+            if (SchemaNameExist(name))
+            {
+                return Json("A Metadata structure with this name already exist. Please choose a other name.", JsonRequestBehavior.AllowGet);
+            }
+
+
             TaskManager = (ImportMetadataStructureTaskManager)Session["TaskManager"];
 
             if (TaskManager.Bus.ContainsKey(ImportMetadataStructureTaskManager.SCHEMA_NAME))
@@ -95,27 +116,32 @@ namespace BExIS.Web.Shell.Areas.DCM.Controllers
             else
                 TaskManager.Bus.Add(ImportMetadataStructureTaskManager.SCHEMA_NAME, name);
 
-            return Content("");
+
+            return Json(true, JsonRequestBehavior.AllowGet);
         }
 
         public bool SchemaNameExist(string SchemaName)
         {
-            MetadataStructureManager msm = new MetadataStructureManager();
 
-            if (msm.Repo.Get().Where(m => m.Name.ToLower().Equals(SchemaName.ToLower())).Count() == 0)
+            using (MetadataStructureManager msm = new MetadataStructureManager())
             {
-                return false;
-            }
-            else
-            {
-                return true;
+
+                if (msm.Repo.Get().Where(m => m.Name.ToLower().Equals(SchemaName.ToLower())).Count() > 0)
+                {
+                    return true;
+                }
+                else
+                {
+                    return false;
+                }
             }
         }
 
 
         public ActionResult GenerateMS()
         {
-          
+            //open schema
+            XmlSchemaManager xmlSchemaManager = new XmlSchemaManager();
 
             string root = "";
             string schemaName = "";
@@ -129,11 +155,6 @@ namespace BExIS.Web.Shell.Areas.DCM.Controllers
             if (TaskManager.Bus.ContainsKey(ImportMetadataStructureTaskManager.SCHEMA_NAME))
                 schemaName = TaskManager.Bus[ImportMetadataStructureTaskManager.SCHEMA_NAME].ToString();
 
-
-           
-
-            TaskManager = (ImportMetadataStructureTaskManager)Session["TaskManager"];
-
             string path = TaskManager.Bus[ImportMetadataStructureTaskManager.FILEPATH].ToString();
             //path = @"https://code.ecoinformatics.org/code/eml/tags/RELEASE_EML_2_1_1/eml.xsd";
 
@@ -141,54 +162,84 @@ namespace BExIS.Web.Shell.Areas.DCM.Controllers
             model.SchemaName = schemaName;
             model.RootNode = root;
 
-            //
-
-            if (SchemaNameExist(schemaName))
+            if (!RegExHelper.IsFilenameValid(schemaName))
             {
-               model.ErrorList.Add(new Error(ErrorType.Other, "A Metadata structure with this name already exist. Please choose a other name."));
+                model.ErrorList.Add(new Error(ErrorType.Other,
+                        "Name : \" " + schemaName + " \" is invalid. These special characters are not allowed : \\/:*?\"<>|"));
             }
-
-            if (String.IsNullOrEmpty(schemaName))
+            else
             {
-                model.ErrorList.Add(new Error(ErrorType.Other, "A Metadata structure must have a name."));
-            }
 
 
-            //open schema
-            XmlSchemaManager xmlSchemaManager = new XmlSchemaManager();
-            xmlSchemaManager.Load(path, GetUserNameOrDefault());
-
-            if (model.ErrorList.Count == 0)
-            {
                 try
                 {
-                    metadataStructureid = xmlSchemaManager.GenerateMetadataStructure(root, schemaName);
+                    //file.WriteLine("check schema exist");
+                    if (SchemaNameExist(schemaName))
+                    {
+                        model.ErrorList.Add(new Error(ErrorType.Other,
+                            "A Metadata structure with this name already exist. Please choose a other name."));
+                    }
+                    else
+                    if (String.IsNullOrEmpty(schemaName))
+                    {
+                        model.ErrorList.Add(new Error(ErrorType.Other, "A Metadata structure must have a name."));
+                    }
+                    else
+                        xmlSchemaManager.Load(path, GetUserNameOrDefault());
+
+                    if (!String.IsNullOrEmpty(model.RootNode) && !xmlSchemaManager.Elements.Any(e => e.Name.Equals(model.RootNode)))
+                    {
+                        model.ErrorList.Add(new Error(ErrorType.Other, "Root node not exist"));
+                    }
+
                 }
                 catch (Exception ex)
                 {
-                    xmlSchemaManager.Delete(schemaName);
                     ModelState.AddModelError("", ex.Message);
                     model.ErrorList.Add(new Error(ErrorType.Other, "Can not create metadatastructure."));
+                    throw ex;
+                }
+
+                if (model.ErrorList.Count == 0)
+                {
+                    try
+                    {
+                        metadataStructureid = xmlSchemaManager.GenerateMetadataStructure(root, schemaName);
+                    }
+                    catch (Exception ex)
+                    {
+                        xmlSchemaManager.Delete(schemaName);
+                        ModelState.AddModelError("", ex.Message);
+                        model.ErrorList.Add(new Error(ErrorType.Other, "Can not create metadatastructure."));
+                    }
+                }
+
+                TaskManager.AddToBus(ImportMetadataStructureTaskManager.MAPPING_FILE_NAME_IMPORT,
+                    xmlSchemaManager.mappingFileNameImport);
+                TaskManager.AddToBus(ImportMetadataStructureTaskManager.MAPPING_FILE_NAME_EXPORT,
+                    xmlSchemaManager.mappingFileNameExport);
+
+
+                model.StepInfo.notExecuted = false;
+
+                if (model.ErrorList.Count == 0)
+                {
+                    model.IsGenerated = true;
+
+                    if (TaskManager.Bus.ContainsKey(ImportMetadataStructureTaskManager.IS_GENERATE))
+                        TaskManager.Bus[ImportMetadataStructureTaskManager.IS_GENERATE] = true;
+                    else
+                        TaskManager.Bus.Add(ImportMetadataStructureTaskManager.IS_GENERATE, true);
+
+                    if (TaskManager.Bus.ContainsKey(ImportMetadataStructureTaskManager.METADATASTRUCTURE_ID))
+                        TaskManager.Bus[ImportMetadataStructureTaskManager.METADATASTRUCTURE_ID] = metadataStructureid;
+                    else
+                        TaskManager.Bus.Add(ImportMetadataStructureTaskManager.METADATASTRUCTURE_ID, metadataStructureid);
                 }
             }
 
-            model.StepInfo.notExecuted = false;
+            return PartialView("ReadSource", model);
 
-            if (model.ErrorList.Count == 0)
-            {
-                if (TaskManager.Bus.ContainsKey(ImportMetadataStructureTaskManager.IS_GENERATE))
-                    TaskManager.Bus[ImportMetadataStructureTaskManager.IS_GENERATE] = true;
-                else
-                    TaskManager.Bus.Add(ImportMetadataStructureTaskManager.IS_GENERATE, true);
-
-                if (TaskManager.Bus.ContainsKey(ImportMetadataStructureTaskManager.METADATASTRUCTURE_ID))
-                    TaskManager.Bus[ImportMetadataStructureTaskManager.METADATASTRUCTURE_ID] = metadataStructureid;
-                else
-                    TaskManager.Bus.Add(ImportMetadataStructureTaskManager.METADATASTRUCTURE_ID, metadataStructureid);
-            }    
-
-
-            return PartialView("ReadSource",model);
         }
 
         // chekc if user exist
